[MIG] delivery_price_method: Migration to 18.0
